Zmiany aby sadzenie i zbieranie działało

+++ Projekt_Farma/Obserwator/Obserwator.java
@@ -4,6 +4,6 @@ import java.io.File;
 
 public interface Obserwator {
 
-    public void aktualizacja(String s, boolean b, int i, File f);
+    public void aktualizacja(String s, boolean b, long i, File f);
 
 }
